Refactor user preferences service to take average embeddings
