+++ backend/app/api/recommendations_api.py
@@ -1,8 +1,7 @@
 # app/api/recommendations_api.py
-from flask import Blueprint, jsonify, request
+from flask import Blueprint, request
 from flask_jwt_extended import jwt_required, get_jwt_identity
 from app.services.recommendation_service import get_recommendations
-from app.services.user_service import get_user_by_id
 
 recommendations_bp = Blueprint('recommendations_bp', __name__)
