+++ backend/app/services/user_preference_service_autoencoder.py
@@ -1,0 +1,44 @@
+# services/user_preferences_service.py
+
+import torch
+import torch.nn as nn
+import torch.optim as optim
+
+class AutoEncoder(nn.Module):
+    def __init__(self, input_dim, hidden_dim):
+        super(AutoEncoder, self).__init__()
+        self.encoder = nn.Sequential(
+            nn.Linear(input_dim, hidden_dim),
+            nn.ReLU()
+        )
+        self.decoder = nn.Sequential(
+            nn.Linear(hidden_dim, input_dim),
+            nn.ReLU()
+        )
+
+    def forward(self, x):
+        encoded = self.encoder(x)
+        decoded = self.decoder(encoded)
+        return decoded, encoded
+
+class UserPreferencesService:
+    def __init__(self, input_dim, hidden_dim):
+        self.model = AutoEncoder(input_dim, hidden_dim)
+        self.criterion = nn.MSELoss()
+        self.optimizer = optim.Adam(self.model.parameters(), lr=0.001)
+
+    def train_autoencoder(self, data_loader, num_epochs=50):
+        for epoch in range(num_epochs):
+            for data in data_loader:
+                inputs = data
+                self.optimizer.zero_grad()
+                outputs, _ = self.model(inputs)
+                loss = self.criterion(outputs, inputs)
+                loss.backward()
+                self.optimizer.step()
+            print(f'Epoch {epoch+1}/{num_epochs}, Loss: {loss.item()}')
+
+    def get_user_embedding(self, user_interactions):
+        with torch.no_grad():
+            _, encoded = self.model(user_interactions)
+        return encoded
